Add files via upload
I created a Word guessing game as my week 2 project at kadaHive, this is how it works, there is already a list made of 100 different words. Which will be randomly selected, one out the list for the user to guess with only the first and last index will be shown to the user.
A simple conversion system to convert from fahrenheit to celsuis.